added legacy yml and seperated classes
