+++ _config/legacy.yml
@@ -1,0 +1,14 @@
+---
+name: gridfieldaddons-legacy
+---
+SilverStripe\ORM\DatabaseAdmin:
+  classname_value_remapping:
+  GridFieldRecordHighlighter: SilverStripe\GridFieldAddOns\GridFieldRecordHighlighter
+  GridFieldUserColumns: SilverStripe\GridFieldAddOns\GridFieldUserColumns
+  GridFieldConfig_ExtendedRecordEditor: SilverStripe\GridFieldAddOns\GridFieldConfig_ExtendedRecordEditor
+  GridFieldConfig_ExtendedRelationEditor: SilverStripe\GridFieldAddOns\GridFieldConfig_ExtendedRelationEditor
+  GridFieldEditableCells: SilverStripe\GridFieldAddOns\GridFieldEditableCells
+  GridFieldUserColumnsExtension: SilverStripe\GridFieldAddOns\GridFieldUserColumnsExtension
+  GridFieldExpandableForm: SilverStripe\GridFieldAddOns\GridFieldExpandableForm
+  GridFieldExpandableForm_ItemRequest: SilverStripe\GridFieldAddOns\GridFieldExpandableForm_ItemRequest
+
