+++ src/GridFieldConfig_ExtendedRelationEditor.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SilverStripe\GridFieldAddOns;
+
+use SilverStripe\GridFieldAddOns\GridFieldUserColumns;
+use SilverStripe\Forms\GridField\GridFieldConfig_RelationEditor;
+
+class GridFieldConfig_ExtendedRelationEditor extends GridFieldConfig_RelationEditor {
+
+	function __construct($itemsPerPage = null) {
+		parent::__construct($itemsPerPage);
+		$this->addComponent(new GridFieldUserColumns());
+	}
+}
